requirements.txt
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
fastapi==0.85.0
uvicorn==0.20.0
numpy==1.23.5
